Add cart app, with urls and views
