+++ help_my_plants/settings.py
@@ -47,6 +47,7 @@ INSTALLED_APPS = [
     'allauth.socialaccount',
     'home',
     'services',
+    'cart',
     'checkout',
     'articles',
