+++ help_my_plants/urls.py
@@ -23,6 +23,7 @@ urlpatterns = [
     path('accounts/', include('allauth.urls')),
     path('', include('home.urls')),
     path('services/', include('services.urls')),
+    path('cart/', include('cart.urls')),
     path('checkout/', include('checkout.urls')),
     path('articles/', include('articles.urls')),
 ] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
